requirements.txt
sanic
sanic-cors
sanic-openapi
boto3
bcrypt
validate_email
py3dns
PyJWT
pymysql
pyyaml
